Updates to the concept engine
A little bit of work to load the textures, also added a common
tgaInfo to Color[][] function, and removed the annoying pointer.
Added a surface to the engine to test the Color[][]
Idea: Start the logo, do a simple 3D map and rotate around it,
then move to the main menu. This way we have tested all the
functionality in a single startup. Static, 3D movement, HUD